create new namespace for cli command
