+++ apps/cli/src/cli-controller/index.ts
@@ -2,10 +2,11 @@ import { Command } from 'commander';
 
 import type { Context } from './types.js';
 import { addSecretCommands } from './secrets.js';
+import { addE2eCommands } from './e2e.js';
 
 export const CliController = (ctx: Context) => {
   const cli = new Command().description(
-    'CLI to interact with the ATJ workspace'
+    'CLI to interact with the Forms workspace'
   );
 
   cli
@@ -16,6 +17,7 @@ export const CliController = (ctx: Context) => {
     });
 
   addSecretCommands(ctx, cli);
+  addE2eCommands(ctx, cli);
 
   return cli;
 };
