+++ e2e/README.md
@@ -8,7 +8,7 @@ To run the tests, run `pnpm test` in this directory.
 In order to test the authenticated state, you can generate an auth session and the associated cookie. To authenticate for these tests, you can run the following command:
 
 ```bash
-pnpm --filter=@gsa-tts/forms-cli cli secrets generate-test-db-session -p ../../packages/server/src/main.db -o ../../e2e/.env
+pnpm --filter=@gsa-tts/forms-cli cli e2e create-test-session -p ../../packages/server/src/main.db -o ../../e2e/.env
 ```
 
 ## Developing tests
