+++ e2e/package.json
@@ -3,7 +3,7 @@
   "version": "0.1.3",
   "private": true,
   "scripts": {
-    "auth": "pnpm --filter=@gsa-tts/forms-cli cli secrets generate-test-db-session -p ../../packages/server/src/main.db -o ../../e2e/.env",
+    "auth": "pnpm --filter=@gsa-tts/forms-cli cli e2e create-test-session -p ../../packages/server/src/main.db -o ../../e2e/.env",
     "dev": "tsc -w",
     "test:e2e:ci": "pnpm auth && pnpm playwright test --headed",
     "test:e2e:dev": "pnpm auth && pnpm playwright test --ui-port=8080 --ui-host=0.0.0.0"
